Rename condition to ternary
